[Redis-connect] Refactor: :rainbow: app.js to app.ts and callback func to promise
app.js를 타입스크립트로 리팩토링 하였고, callback 함수로 이루어진 비동기처리 로직을 promise로 사용할 수 있게끔 하는 방법을 추가하였다.
